#300 moved function to utils file, added tests
